+++ tests/acceptance/edit-form-validation-test/validation-file-test.js
@@ -17,8 +17,10 @@ executeTest('check operation file', (store, assert, app) => {
     // Check default validationmessage text.
     assert.equal($validationFlexberryErrorLable.text().trim(), 'File is required', 'Flexberry file have default value');
 
-    let $validationFlexberryLookupButtons = $('.ui.button');
-    let $validationFlexberryLookupButton = $($validationFlexberryLookupButtons[2]);
+    let $validationFlexberryLookup = $('.flexberry-lookup');
+    let $validationFlexberryLookupInput = $validationFlexberryLookup.children('.input');
+    let $validationFlexberryLookupButton = $validationFlexberryLookupInput.children('.ui-change.button');
+
 
     // Click lookup button.
     run(() => {